Create Printing and For Looping Through Dictionaries and Nested Dictionaries
In this example, we have created and empty dictionary and a country capital list. We created the country capital dictionary with and without a function.



We printed the results which yielded the output in a horizontal manner. For 3 key pair values, it may not be cumbersome to read the output but what if we have 10 key pair values?



To solve this problem, we will use For Loops in Lists, Sets, Tuples and Dictionaries!



For Loops will enable the result to be printed in a vertical format which is much more easier to read.