Create Password_Strength_Checker_NestedFunctions
Password Safety Checker Project

For this exercise, we will check the eligibility of a password strength based on the combination of number of characters of the password (8), upper case, lower case, special symbols and numbers. 

If the user input password has a combination of minimum character of 8, presence of small letter alphabets, capital letter alphabets, numbers and special symbols, then the statement that password strength is good will be printed! 

#Functions
#UserInputs
#NestedFunctions
#ReturnStatements
#ForLoops
#ConditionalStatements
#LogicalOperators
#PrintStatements